Ksyao2002/permissionless tranches v2

+++ contracts/misc/interfaces/IWETHGateway.sol
@@ -4,21 +4,21 @@ pragma solidity >=0.8.0;
 interface IWETHGateway {
     function depositETH(
         address lendingPool,
-        uint8 trancheId,
+        uint64 trancheId,
         address onBehalfOf,
         uint16 referralCode
     ) external payable;
 
     function withdrawETH(
         address lendingPool,
-        uint8 trancheId,
+        uint64 trancheId,
         uint256 amount,
         address onBehalfOf
     ) external;
 
     function repayETH(
         address lendingPool,
-        uint8 trancheId,
+        uint64 trancheId,
         uint256 amount,
         uint256 rateMode,
         address onBehalfOf
@@ -26,7 +26,7 @@ interface IWETHGateway {
 
     function borrowETH(
         address lendingPool,
-        uint8 trancheId,
+        uint64 trancheId,
         uint256 amount,
         uint256 interesRateMode,
         uint16 referralCode
